+++ tests/test_calibration.py
@@ -17,6 +17,8 @@ from extra.calibration import (
     JUNGFRAUConditions,
     LPDConditions,
     SingleConstant,
+    DetectorData,
+    PhysicalDetectorUnit
 )
 
 # Most of these tests use saved HTTP responses by default (with pytest-recording).
@@ -376,3 +378,78 @@ def test_conditions_markdown():
     cond = LPDConditions()
     md = MarkdownFormatter()(cond)  # Smoketest
     assert isinstance(md, str)
+
+
+@pytest.mark.vcr
+def test_DetectorData_from_identifier():
+    # SPB-AGIPD, multi-module detector with full CalCat entries
+    agipd = DetectorData.from_identifier('SPB_DET_AGIPD1M-1')
+    repr(agipd)
+
+    assert agipd
+    assert agipd.identifier == 'SPB_DET_AGIPD1M-1'
+    assert agipd.source_names[0] == 'SPB_DET_AGIPD1M-1/DET/0CH0:xtdf'
+    assert agipd.pdu_detector_types == {'AGIPD-Type'}
+    assert agipd.detector_type == 'AGIPD-Type'
+    assert len(agipd) == agipd.number_of_modules
+    assert list(agipd)[0] == next(iter(agipd.keys())) == 'AGIPD00'
+    assert isinstance(next(iter(agipd.values())), PhysicalDetectorUnit)
+    assert agipd[0] == agipd['AGIPD00']
+
+    # PDU
+    pdu = agipd[0]
+    assert pdu.aggregator == 'AGIPD00'
+    assert pdu.ccv_params == (
+        'AGIPD_SIV1_AGIPDV11_M517', 101003000000, 'AGIPD-Type')
+
+    # SQS-pnCCD, single-module detector with partial CalCat entries
+    pnccd = DetectorData.from_identifier('SQS_NQS_PNCCD1MP')
+    repr(pnccd)
+
+    assert pnccd
+    assert len(pnccd) == 1
+    assert pnccd.number_of_modules is None
+
+    with pytest.raises(AssertionError):
+        pnccd.first_module_index
+
+    # SQS-DSSC, PDU-less detector
+    dssc = DetectorData.from_identifier('SQS_DET_DSSC1M-1')
+    repr(dssc)
+
+    assert not dssc
+    assert len(dssc) == 0
+    assert not dssc.pdu_detector_types
+
+    with pytest.raises(ValueError):
+        dssc.detector_type
+
+
+@pytest.mark.vcr
+def test_DetectorData_from_instrument():
+    with pytest.raises(ValueError):
+        DetectorData.from_instrument('SPB')  # More than one detector.
+
+    with pytest.raises(ValueError):
+        DetectorData.from_instrument('SPB', '*JF*')  # More than one JF.
+
+    # Sufficiently narrow glob.
+    jf4m = DetectorData.from_instrument('SPB', '*JF4M')
+    assert jf4m.identifier == 'SPB_IRDA_JF4M'
+
+    # Instrument with single detector.
+    hirex = DetectorData.from_instrument('SA1')
+    assert hirex.identifier == 'SA1_XTD9_HIREX'
+
+
+@pytest.mark.vcr
+def test_DetectorData_list_by_instrument():
+    assert DetectorData.list_by_instrument('SCS') == [
+        'SCS_DET_DSSC1M-1', 'SCS_HRIXS_JUNGF', 'SCS_XOX_GH21',
+        'SCS_XOX_GH22', 'SCS_DET_DSSC2']
+
+
+@pytest.mark.vcr
+def test_DetectorData_from_CalibrationData():
+    agipd_cd = CalibrationData.from_report(3757)
+    assert agipd_cd.detector().identifier == 'SPB_DET_AGIPD1M-1'